requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-telegram-bot==20.7
twilio==8.10.0
elevenlabs==0.2.26
python-dotenv==1.0.0
gunicorn==21.2.0
pyngrok==7.0.0
asyncio
aiohttp==3.9.1

# OpenTelemetry tracing (optional)
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-flask>=0.37b0
opentelemetry-instrumentation-requests>=0.37b0
opentelemetry-exporter-otlp>=1.12.0